+++ packages/account-abstraction/src/SmartAccountAPI.ts
@@ -1,11 +1,11 @@
 import { BigNumber, BigNumberish } from 'ethers'
-// import { EntryPointContractV100 } from '@biconomy/ethers-lib'
+// import { EntryPointContractV100 } from '@biconomy-devx/ethers-lib'
 import { EntryPoint } from '@account-abstraction/contracts'
 import { ClientConfig } from './ClientConfig' // added in this design
 import { arrayify, hexConcat } from 'ethers/lib/utils'
 import { Signer } from '@ethersproject/abstract-signer'
 import { TransactionDetailsForBatchUserOp } from './TransactionDetailsForUserOp'
-import { UserOperation } from '@biconomy/core-types'
+import { UserOperation } from '@biconomy-devx/core-types'
 import { BaseApiParams, BaseAccountAPI } from './BaseAccountAPI'
 import { Provider } from '@ethersproject/providers'
 import { BiconomyPaymasterAPI } from './BiconomyPaymasterAPI'
@@ -15,7 +15,7 @@ import {
   Logger,
   deployCounterFactualEncodedData,
   EIP1559_UNSUPPORTED_NETWORKS
-} from '@biconomy/common'
+} from '@biconomy-devx/common'
 import { ethers } from 'ethers/lib'
 
 // may use...