+++ packages/account-abstraction/src/calcPreVerificationGas.ts
@@ -1,5 +1,5 @@
-import { UserOperation } from '@biconomy/core-types' // review
-import { NotPromise, packUserOp } from '@biconomy/common' // '@account-abstraction/utils'
+import { UserOperation } from '@biconomy-devx/core-types' // review
+import { NotPromise, packUserOp } from '@biconomy-devx/common' // '@account-abstraction/utils'
 import { arrayify, hexlify } from 'ethers/lib/utils'
 
 export interface GasOverheads {