+++ packages/account-abstraction/test/1-SmartAccountAPI.test.ts
@@ -3,9 +3,9 @@ import { Wallet } from 'ethers'
 import { parseEther } from 'ethers/lib/utils'
 import { expect } from 'chai'
 import { ethers } from 'hardhat'
-import { UserOperation } from '@biconomy/core-types'
+import { UserOperation } from '@biconomy-devx/core-types'
 import { SmartAccountAPI } from '../src'
-import { SampleRecipient, SampleRecipient__factory } from '@biconomy/common/dist/src/types'
+import { SampleRecipient, SampleRecipient__factory } from '@biconomy-devx/common/dist/src/types'
 
 import {
   SmartWalletContractFactoryV100,
@@ -14,7 +14,7 @@ import {
   SmartWalletFactoryContractV100,
   EntryPointContractV100,
   SmartWalletContractV100
-} from '@biconomy/ethers-lib'
+} from '@biconomy-devx/ethers-lib'
 
 const provider = ethers.provider
 const signer = provider.getSigner()