+++ packages/account-abstraction/test/2-ERC4337EthersSigner.test.ts
@@ -1,4 +1,4 @@
-import { SampleRecipient, SampleRecipient__factory } from '@biconomy/common/dist/src/types'
+import { SampleRecipient, SampleRecipient__factory } from '@biconomy-devx/common/dist/src/types'
 import { ethers } from 'hardhat'
 import { ClientConfig, ERC4337EthersProvider, newProvider } from '../src'
 import { EntryPoint, EntryPoint__factory } from '@account-abstraction/contracts'
@@ -12,7 +12,7 @@ import {
   SmartWalletFactoryContractV100,
   EntryPointContractV100,
   SmartWalletContractV100
-} from '@biconomy/ethers-lib'
+} from '@biconomy-devx/ethers-lib'
 
 const provider = ethers.provider
 const signer = provider.getSigner()