+++ packages/ethers-lib/src/contracts/SmartWallet/v1.0.0/SmartWalletContractEthers.ts
@@ -6,7 +6,7 @@ import {
   ExecTransaction,
   IFeeRefundV1_0_1,
   ITransactionResult
-} from '@biconomy/core-types'
+} from '@biconomy-devx/core-types'
 import { toTxResult } from '../../../utils'
 import { SmartWalletContractV100 as SmartWalletContract_TypeChain } from '../../../../typechain/src/ethers-v5/v1.0.0/SmartWalletContractV100'
 import { SmartWalletContractV100Interface } from '../../../../typechain/src/ethers-v5/v1.0.0/SmartWalletContractV100'