+++ packages/transactions/CHANGELOG.md
@@ -22,4 +22,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ## 1.0.0 (2023-01-03)
 
 
-**Note:** Version bump only for package @biconomy/transactions
+**Note:** Version bump only for package @biconomy-devx/transactions